User authentication and permissions

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -1,12 +1,23 @@
 const bcrypt = require('bcryptjs');
 const jwt = require('jsonwebtoken');
+const { randomBytes } = require('crypto');
+const { promisify } = require('util');
+const { transport, makeANiceEmail } = require('../mail');
+const { hasPermission } = require('../utils');
 
 const Mutations = {
   async createItem(parent, args, ctx, info) {
-    // TODO: check if they are logged in 
+    if(!ctx.request.userId) {
+      throw new Error('You must be logged in to do that');
+    }
 
     const item = await ctx.db.mutation.createItem({
       data: {
+        user: {
+          connect: {
+            id: ctx.request.userId
+          }
+        },
         ...args
       }
     }, info);
@@ -31,10 +42,15 @@ const Mutations = {
     const where = { id: args.id };
 
     // 1. find the item
-    const item = await ctx.db.query.item({ where }, `{ id, title}`);
+    const item = await ctx.db.query.item({ where }, `{ id title user { id }}`);
 
     // 2. check if they own that item / have permissions
-    // TODO
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission => ['ADMIN', 'ITEMDELETE'].includes(permission));
+
+    if(!ownsItem && !hasPermissions) {
+      throw new Error('You don\'t have permission to do that');
+    }
 
     // 3. delete it
     return ctx.db.mutation.deleteItem({ where }, info);
@@ -67,6 +83,139 @@ const Mutations = {
 
     // return the user to the browser
     return user;
+  },
+
+  async signin(parent, {email, password}, ctx, info) {
+    // check if there is a user with that email 
+    const user = await ctx.db.query.user({ where: { email: email }});
+    if(!user) {
+      throw new Error(`No such user found for email ${email}`);
+    }
+
+    // check if their password is correct
+    const valid = await bcrypt.compare(password, user.password);
+    if(!valid) {
+      throw new Error('Invalid password');
+    }
+
+    // generate JWT token
+    const token = jwt.sign({ userId: user.id }, process.env.APP_SECRET);
+
+    // set the cookie with the token
+    ctx.response.cookie('token', token, {
+      httpOnly: true,
+      maxAge: 1000 * 60 * 60 * 24 * 365, // 1 year cookie 
+    });
+
+    // return the user
+    return user;
+  },
+
+  signout(parent, args, ctx, info) {
+    ctx.response.clearCookie('token');
+    return { message: 'Goodbye!' };
+  },
+
+  async requestReset(parent, args, ctx, info) {
+    // check if this is a real user
+    const user = await ctx.db.query.user({ where: { email: args.email } });
+    if(!user) {
+      throw new Error(`No such user found for email ${args.email}`);
+    }
+
+    // set a reset token and expiry on that user
+    const randomBytesPromisified = promisify(randomBytes);
+    const resetToken = (await randomBytesPromisified(20)).toString('hex');
+    const resetTokenExpiry = Date.now() + 3600000; // 1 hour from now
+    const res = await ctx.db.mutation.updateUser({
+      where: { email: args.email },
+      data: { resetToken, resetTokenExpiry }
+    });
+    
+    // email them that reset token 
+    const mailRes = await transport.sendMail({
+      from: 'hello@staceyfenton.com',
+      to: user.email,
+      subject: 'Your password reset token',
+      html: makeANiceEmail(`Your password reset token is here! \n\n 
+      <a href="${process.env.FRONTEND_URL}/reset?resetToken=${resetToken}">Reset your password</a>.`)
+    });
+
+    // return the message
+    return { message: 'Thanks!' }
+  }, 
+
+  async resetPassword(parent, args, ctx, info) {
+    // check if the passwords match
+    if(args.password !== args.confirmPassword) {
+      throw new Error('Your passwords don\'t match');
+    }
+    // check if its a legit reset token
+
+    // check if it has expired
+    const [user] = await ctx.db.query.users({
+      where: {
+        resetToken: args.resetToken,
+        resetTokenExpiry_gte: Date.now() - 3600000
+      }
+    });
+    if(!user) {
+      throw new Error('This token is either invalid or expired');
+    }
+
+    // hash their new password
+    const password = await bcrypt.hash(args.password, 10);
+
+    // save the new password to the user and remove old resetToken fields
+    const updatedUser = await ctx.db.mutation.updateUser({
+      where: { email: user.email }, 
+      data: {
+        password,
+        resetToken: null,
+        resetTokenExpiry: null
+      }
+    });
+
+    // generate JWT
+    const token = jwt.sign({ userId: updatedUser.id }, process.env.APP_SECRET);
+
+    // set the JWT cookie
+    ctx.response.cookie('token', token, {
+      httpOnly: true,
+      maxAge: 1000 * 60 * 60 * 24 * 365, // 1 year cookie 
+    });
+
+    // return updated user
+    return updatedUser; 
+  },
+
+  async updatePermissions(parent, args, ctx, info) {
+    // check if they are logged in
+    if(!ctx.request.userId) {
+      throw new Error('You must be logged in');
+    }
+
+    // query the current user
+    const currentUser = await ctx.db.query.user({
+      where: {
+        id: ctx.request.userId
+      }
+    }, info);
+
+    // check if they have permissions to do this
+    hasPermission(currentUser, ['ADMIN', 'PERMISSIONUPDATE']);
+
+    // update the permissions
+    return ctx.db.mutation.updateUser({
+      data: {
+        permissions: {
+          set: args.permissions
+        }
+      },
+      where: {
+        id: args.userId
+      }
+    }, info);
   }
 };
